Revert "🔨 refactor: `src.core.accido.edge_cases` -> `src.core.accido._edge_cases`"
This reverts commit 30cd1d5840d51cfbeed57dd617923c83080648b4.
