+++ src/core/accido/_class_noun.py
@@ -7,7 +7,7 @@ from functools import total_ordering
 from typing import TYPE_CHECKING, overload
 
 from ._class_word import _Word
-from ._edge_cases import IRREGULAR_NOUNS
+from .edge_cases import IRREGULAR_NOUNS
 from .exceptions import InvalidInputError
 from .misc import (
     Case,
