+++ src/core/transfero/_adjective_inflection.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 import lemminflect
 
 from ..accido.misc import ComponentsType, Degree
-from ._edge_cases import NOT_COMPARABLE_ADJECTIVES
+from .edge_cases import NOT_COMPARABLE_ADJECTIVES
 from .exceptions import InvalidComponentsError, InvalidWordError
 
 if TYPE_CHECKING:
