+++ tests/accido_test/pronoun_test.py
@@ -4,7 +4,7 @@ import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 
 import pytest
-from src.core.accido._edge_cases import PRONOUNS
+from src.core.accido.edge_cases import PRONOUNS
 from src.core.accido.endings import Noun, Pronoun
 from src.core.accido.exceptions import InvalidInputError
 from src.core.accido.misc import Case, EndingComponents, Gender, MultipleMeanings, Number
